refactor: move filename and filetype out of Buffer into View
